test: update revenue source integ tests
